Configure Rubocop
This is an opinionated configuration which you of course are at liberty
to amend.

Here are some of the reasons behind these choices:

```
Layout/LineLength:
  Max: 125
```

I defer to Linus Torvalds: https://lkml.org/lkml/2020/5/29/1038

```
Metrics/AbcSize:
  Exclude:
    - db/migrate/**/*
```

The generated migrations sometimes breach this limit. I see no sense in
making people refactor generated code.

```
Metrics/BlockLength:
  Exclude:
    - spec/**/*
```

The structure of RSpec blocks makes this cop impossible to comply with
for specs.

```
Metrics/MethodLength:
  Exclude:
    - db/migrate/**/*
```

This is also difficult to comply with for even the simplest and most
orderly of specs.

```
RSpec/ExampleLength:
  Max: 10
```

Similar to `Metrics/MethodLength`, this is impossible to comply with in
real life.

```
Style/Documentation:
  Enabled: false
```

Comments in code are just as likely to mislead as to illuminate.

```
Style/TrailingCommaInArrayLiteral:
  EnforcedStyleForMultiline: comma

Style/TrailingCommaInArguments:
  EnforcedStyleForMultiline: comma

Style/TrailingCommaInHashLiteral:
  EnforcedStyleForMultiline: comma
```

I know it's ugly. I know. But you'll thank me when you're reviewing
code because the diffs are so much more logical.